fix: fix environment validation
